Rework GetSqlWCharSize for Mac
