+++ flightsql-odbc/flightsql-odbc/flightsql-odbc-clone/flightsql-odbc/odbcabstraction/include/odbcabstraction/encoding.h
@@ -6,7 +6,6 @@
 
 #pragma once
 
-#include <boost/algorithm/string/predicate.hpp>
 #include <odbcabstraction/exceptions.h>
 #include <cassert>
 #include <codecvt>
@@ -15,47 +14,20 @@
 #include <vector>
 
 #if defined(__APPLE__)
-#include <dlfcn.h>
-#include <mutex>
+#include <atomic>
 #endif
 
 namespace driver {
 namespace odbcabstraction {
 
 #if defined(__APPLE__)
-namespace {
-static std::mutex SqlWCharSizeMutex;
 static std::atomic<size_t> SqlWCharSize{0};
 
-bool IsUsingIODBC() {
-  // Detects iODBC by looking up by symbol iodbc_version
-  void* handle = dlsym(RTLD_DEFAULT, "iodbc_version");
-  bool using_iodbc = handle != nullptr;
-  dlclose(handle);
-
-  return using_iodbc;
-}
-}
+void ComputeSqlWCharSize();
 
 inline size_t GetSqlWCharSize() {
   if (SqlWCharSize == 0) {
-    std::unique_lock<std::mutex> lock(SqlWCharSizeMutex);
-    if (SqlWCharSize == 0) { // double-checked locking
-
-      const char *env_p = std::getenv("WCHAR_ENCODING");
-      if (env_p) {
-        if (boost::iequals(env_p, "UTF-16")) {
-          SqlWCharSize = sizeof(char16_t);
-          return SqlWCharSize;
-        } else if (boost::iequals(env_p, "UTF-32")) {
-          SqlWCharSize = sizeof(char32_t);
-          return SqlWCharSize;
-        }
-      }
-
-      bool using_iodbc = IsUsingIODBC();
-      SqlWCharSize = using_iodbc ? sizeof(char32_t) : sizeof(char16_t);
-    }
+    ComputeSqlWCharSize();
   }
 
   return SqlWCharSize;
